docs: allow to run styleguide with legacy theme
